moved AddFacility from containers to components
